Memory changes

+++ TurtleGames.VoxelEngine/ChunkData.cs
@@ -6,8 +6,8 @@ public class ChunkData
 {
     public ChunkVector Position { get; set; }
     public Vector2 Size { get; set; }
-    public int[,,] Chunk { get; set; }
-    public int Height { get; set; }
+    public ushort[,,] Chunk { get; set; }
+    public ushort Height { get; set; }
     public bool Calculated { get; set; }
 
 }
